+++ crates/monty/src/args/mod.rs
@@ -7,7 +7,7 @@ use std::{mem, slice, vec::IntoIter};
 
 pub(crate) use bind_native::{Bound, ErrorFamily, Param, ParamKind, ParamSpec, bind};
 pub(crate) use bind_python::Signature;
-pub(crate) use from_value::{ArgErrCtx, FromValue, LaxBool, StrArg};
+pub(crate) use from_value::{ArgErrCtx, FromValue, FromValueFail, LaxBool, StrArg, is_long_int};
 pub(crate) use monty_macros::{FromArgs, ToArgs};
 pub(crate) use to_monty_object::ToMontyObject;
